v2201.3 changed GEN and Starter CAS Annunciation
- Update Version info
- The GCU
provides for automatic starter cutoff when engine RPM is above 46%.
Below 46%, the starter-generator functions as a starter, and above
46%, the starter-generator functions as a generator

GENERATOR OFF (AMBER) - Indicates that the generator is not
connected to the electrical bus with engine running.

STARTER ON (AMBER) - Indicates the starter-generator is
operating in starter mode.
After
the engine has stabilized at idle, the amber STARTER ON annunciator
should not be shown. If this annunciator remains ON, it indicates the
starter has not been automatically disengaged during the engine
starting sequence due to a failed speed sensor.

The airplane is equipped with two starter contactors. One is used for
starts using external power and the other for battery starts. If either
contactor does not open after reaching approximately 46% Ng, the
amber STARTER ON annunciator will remain on. In most cases when
this occurs, the generator will not transfer over to the generator mode,
and the amber GENERATOR OFF annunciator will remain on.